backend/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
chunkr-ai
pinecone-client==3.0.0
openai==1.54.0
pyyaml==6.0.1